Refactor Azure Cosmos DB logic
